make project crud in admin

+++ application/models/User_model.php
@@ -53,6 +53,7 @@ class User_model extends CI_Model
 		$this->db->select('*');
 		$this->db->from('user');
 		$this->db->where('email', $email);
+		$this->db->where('user.is_delete', 0);
 		$query = $this->db->get();
 		return $query->row_array();
 	}
@@ -122,8 +123,10 @@ class User_model extends CI_Model
 	 */
 	public function delete_user($user_id)
 	{
-		$this->db->where('id', $user_id);
-		$this->db->delete('user');
+		if ($user_id > 0) {
+			$this->db->where('id', $user_id);
+			$this->db->delete('user');
+		}
 		return $this->db->affected_rows();
 	}
